+++ packages/vendure-plugin-goedgepickt/src/goedgepickt.plugin.ts
@@ -40,7 +40,7 @@ import { customFields } from './api/custom-fields';
 
     return config;
   },
-  compatibility: '^2.0.0',
+  compatibility: '>=2.2.0',
 })
 export class GoedgepicktPlugin {
   static config: GoedgepicktPluginConfig;