+++ packages/vendure-plugin-popularity-scores/src/popularity-scores.plugin.ts
@@ -62,7 +62,7 @@ export interface PopularityScoresPluginConfig {
     });
     return config;
   },
-  compatibility: '^2.0.0',
+  compatibility: '>=2.2.0',
 })
 export class PopularityScoresPlugin {
   static config: PopularityScoresPluginConfig;